Move CompletionConfig to a separate module
